[Canvas][Layout Engine] TypeScript conversion - Stage I (#29390)
* Perf: avoid matrix transpose in canvas layout that served only initial docu purpose, don't cons extra arrays
* Chore: switch to ES2015 import/export and direct function access
* Chore: TS conversion for some of the layout engine files
* Chore: rework TS linting for all files under `aeroelastic` (even the `.js` ones)
