+++ x-pack/plugins/canvas/public/components/alignment_guide/alignment_guide.js
@@ -6,7 +6,7 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import aero from '../../lib/aeroelastic';
+import { toCSS } from '../../lib/aeroelastic';
 
 export const AlignmentGuide = ({ transformMatrix, width, height }) => {
   const newStyle = {
@@ -16,7 +16,7 @@ export const AlignmentGuide = ({ transformMatrix, width, height }) => {
     marginTop: -height / 2,
     background: 'magenta',
     position: 'absolute',
-    transform: aero.dom.matrixToCSS(transformMatrix),
+    transform: toCSS(transformMatrix),
   };
   return (
     <div
