+++ x-pack/plugins/canvas/public/components/border_resize_handle/border_resize_handle.js
@@ -6,12 +6,12 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import aero from '../../lib/aeroelastic';
+import { toCSS } from '../../lib/aeroelastic';
 
 export const BorderResizeHandle = ({ transformMatrix }) => (
   <div
     className="canvasBorderResizeHandle canvasLayoutAnnotation"
-    style={{ transform: aero.dom.matrixToCSS(transformMatrix) }}
+    style={{ transform: toCSS(transformMatrix) }}
   />
 );
