+++ x-pack/plugins/canvas/public/components/rotation_handle/rotation_handle.js
@@ -6,12 +6,12 @@
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import aero from '../../lib/aeroelastic';
+import { toCSS } from '../../lib/aeroelastic';
 
 export const RotationHandle = ({ transformMatrix }) => (
   <div
     className="canvasRotationHandle canvasRotationHandle--connector canvasLayoutAnnotation"
-    style={{ transform: aero.dom.matrixToCSS(transformMatrix) }}
+    style={{ transform: toCSS(transformMatrix) }}
   >
     <div className="canvasRotationHandle--handle" />
   </div>
