+++ x-pack/plugins/canvas/public/lib/aeroelastic/geometry.js
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-const matrix = require('./matrix');
+import { invert, mvMultiply, normalize, ORIGIN } from './matrix';
 
 /**
  * Pure calculations with geometry awareness - a set of rectangles with known size (a, b) and projection (transform matrix)
@@ -40,9 +40,9 @@ const shapesAtPoint = (shapes, x, y) =>
 
     // Determine z (depth) by composing the x, y vector out of local unit x and unit y vectors; by knowing the
     // scalar multipliers for the unit x and unit y vectors, we can determine z from their respective 'slope' (gradient)
-    const centerPoint = matrix.normalize(matrix.mvMultiply(transformMatrix, matrix.ORIGIN));
-    const rightPoint = matrix.normalize(matrix.mvMultiply(transformMatrix, [1, 0, 0, 1]));
-    const upPoint = matrix.normalize(matrix.mvMultiply(transformMatrix, [0, 1, 0, 1]));
+    const centerPoint = normalize(mvMultiply(transformMatrix, ORIGIN));
+    const rightPoint = normalize(mvMultiply(transformMatrix, [1, 0, 0, 1]));
+    const upPoint = normalize(mvMultiply(transformMatrix, [0, 1, 0, 1]));
     const x0 = rightPoint[0] - centerPoint[0];
     const y0 = rightPoint[1] - centerPoint[1];
     const x1 = upPoint[0] - centerPoint[0];
@@ -58,8 +58,8 @@ const shapesAtPoint = (shapes, x, y) =>
     // Hmm maybe we should reuse the above right and up unit vectors to establish whether we're within the (a, b) 'radius'
     // rather than using matrix inversion. Bound to be cheaper.
 
-    const inverseProjection = matrix.invert(transformMatrix);
-    const intersection = matrix.normalize(matrix.mvMultiply(inverseProjection, [x, y, z, 1]));
+    const inverseProjection = invert(transformMatrix);
+    const intersection = normalize(mvMultiply(inverseProjection, [x, y, z, 1]));
     const [sx, sy] = intersection;
 
     // z is needed downstream, to tell which one is the closest shape hit by an x, y ray (shapes can be tilted in z)
@@ -76,18 +76,13 @@ const shapesAtPoint = (shapes, x, y) =>
 //
 // If it were a right handed coordinate system, AND Y still pointed down, then Z should increase away from the
 // viewer. But that's not the case. So we maximize the Z value to tell what's on top.
-const shapesAt = (shapes, { x, y }) =>
+export const shapesAt = (shapes, { x, y }) =>
   shapesAtPoint(shapes, x, y)
     .filter(shape => shape.inside)
     .sort((shape1, shape2) => shape2.z - shape1.z || shape2.index - shape1.index) // stable sort: DOM insertion order!!!
     .map(shape => shape.shape); // decreasing order, ie. from front (closest to viewer) to back
 
-const getExtremum = (transformMatrix, a, b) =>
-  matrix.normalize(matrix.mvMultiply(transformMatrix, [a, b, 0, 1]));
+const getExtremum = (transformMatrix, a, b) => normalize(mvMultiply(transformMatrix, [a, b, 0, 1]));
 
-const landmarkPoint = (a, b, transformMatrix, k, l) => getExtremum(transformMatrix, k * a, l * b);
-
-module.exports = {
-  landmarkPoint,
-  shapesAt,
-};
+export const landmarkPoint = (a, b, transformMatrix, k, l) =>
+  getExtremum(transformMatrix, k * a, l * b);
