+++ x-pack/plugins/canvas/public/lib/aeroelastic/gestures.js
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-const { select, selectReduce } = require('./state');
+import { select, selectReduce } from './state';
 
 // Only needed to shuffle some modifier keys for Apple keyboards as per vector editing software conventions,
 // so it's OK that user agent strings are not reliable; in case it's spoofed, it'll just work with a slightly
@@ -43,15 +43,16 @@ const keyFromMouse = select(({ type, payload: { altKey, metaKey, shiftKey, ctrlK
   type === 'cursorPosition' || type === 'mouseEvent' ? { altKey, metaKey, shiftKey, ctrlKey } : {}
 )(primaryUpdate);
 
-const metaHeld = select(appleKeyboard ? e => e.metaKey : e => e.altKey)(keyFromMouse);
-const optionHeld = select(appleKeyboard ? e => e.altKey : e => e.ctrlKey)(keyFromMouse);
-const shiftHeld = select(e => e.shiftKey)(keyFromMouse);
+export const metaHeld = select(appleKeyboard ? e => e.metaKey : e => e.altKey)(keyFromMouse);
+export const optionHeld = select(appleKeyboard ? e => e.altKey : e => e.ctrlKey)(keyFromMouse);
+export const shiftHeld = select(e => e.shiftKey)(keyFromMouse);
 
-const cursorPosition = selectReduce((previous, position) => position || previous, { x: 0, y: 0 })(
-  rawCursorPosition
-);
+export const cursorPosition = selectReduce((previous, position) => position || previous, {
+  x: 0,
+  y: 0,
+})(rawCursorPosition);
 
-const mouseButton = selectReduce(
+export const mouseButton = selectReduce(
   (prev, next) => {
     if (!next) {
       return prev;
@@ -66,12 +67,12 @@ const mouseButton = selectReduce(
   { down: false, uid: null }
 )(mouseButtonEvent);
 
-const mouseIsDown = selectReduce(
+export const mouseIsDown = selectReduce(
   (previous, next) => (next ? next.event === 'mouseDown' : previous),
   false
 )(mouseButtonEvent);
 
-const gestureEnd = select(
+export const gestureEnd = select(
   action =>
     action &&
     (action.type === 'actionEvent' ||
@@ -85,39 +86,39 @@ const gestureEnd = select(
  *    Edit: http://stable.ascii-flow.appspot.com/#567671116534197027/776257435
  *
  *
- *                             mouseIsDown
- *        initial state: 'up' +-----------> 'downed'
- *                        ^ ^                 +  +
- *                        | |  !mouseIsDown   |  |
- *           !mouseIsDown | +-----------------+  | mouseIsDown && movedAlready
- *                        |                      |
- *                        +----+ 'dragging' <----+
+ *                             mouseNowDown
+ *        initial state: 'up' +------------> 'downed'
+ *                        ^ ^                  +  +
+ *                        | |  !mouseNowDown   |  |
+ *          !mouseNowDown | +------------------+  | mouseNowDown && movedAlready
+ *                        |                       |
+ *                        +----+ 'dragging' <-----+
  *                                +      ^
  *                                |      |
  *                                +------+
- *                               mouseIsDown
+ *                               mouseNowDown
  *
  */
-const mouseButtonStateTransitions = (state, mouseIsDown, movedAlready) => {
+const mouseButtonStateTransitions = (state, mouseNowDown, movedAlready) => {
   switch (state) {
     case 'up':
-      return mouseIsDown ? 'downed' : 'up';
+      return mouseNowDown ? 'downed' : 'up';
     case 'downed':
-      if (mouseIsDown) {
+      if (mouseNowDown) {
         return movedAlready ? 'dragging' : 'downed';
       } else {
         return 'up';
       }
 
     case 'dragging':
-      return mouseIsDown ? 'dragging' : 'up';
+      return mouseNowDown ? 'dragging' : 'up';
   }
 };
 
 const mouseButtonState = selectReduce(
-  ({ buttonState, downX, downY }, mouseIsDown, { x, y }) => {
+  ({ buttonState, downX, downY }, mouseNowDown, { x, y }) => {
     const movedAlready = x !== downX || y !== downY;
-    const newButtonState = mouseButtonStateTransitions(buttonState, mouseIsDown, movedAlready);
+    const newButtonState = mouseButtonStateTransitions(buttonState, mouseNowDown, movedAlready);
     return {
       buttonState: newButtonState,
       downX: newButtonState === 'downed' ? x : downX,
@@ -127,11 +128,11 @@ const mouseButtonState = selectReduce(
   { buttonState: 'up', downX: null, downY: null }
 )(mouseIsDown, cursorPosition);
 
-const mouseDowned = select(state => state.buttonState === 'downed')(mouseButtonState);
+export const mouseDowned = select(state => state.buttonState === 'downed')(mouseButtonState);
 
-const dragging = select(state => state.buttonState === 'dragging')(mouseButtonState);
+export const dragging = select(state => state.buttonState === 'dragging')(mouseButtonState);
 
-const dragVector = select(({ buttonState, downX, downY }, { x, y }) => ({
+export const dragVector = select(({ buttonState, downX, downY }, { x, y }) => ({
   down: buttonState !== 'up',
   x0: downX,
   y0: downY,
@@ -139,20 +140,6 @@ const dragVector = select(({ buttonState, downX, downY }, { x, y }) => ({
   y1: y,
 }))(mouseButtonState, cursorPosition);
 
-const actionEvent = select(action => (action.type === 'actionEvent' ? action.payload : null))(
-  primaryUpdate
-);
-
-module.exports = {
-  actionEvent,
-  dragging,
-  dragVector,
-  cursorPosition,
-  gestureEnd,
-  metaHeld,
-  mouseButton,
-  mouseDowned,
-  mouseIsDown,
-  optionHeld,
-  shiftHeld,
-};
+export const actionEvent = select(action =>
+  action.type === 'actionEvent' ? action.payload : null
+)(primaryUpdate);
