+++ x-pack/plugins/canvas/public/lib/aeroelastic/index.js
@@ -4,18 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-const dom = require('./dom');
-const geometry = require('./geometry');
-const gestures = require('./gestures');
-const layout = require('./layout');
-const matrix = require('./matrix');
-const state = require('./state');
+import { matrixToCSS } from './dom';
+import { nextScene } from './layout';
+import { primaryUpdate } from './layout_functions';
+import { multiply, rotateZ, translate } from './matrix';
+import { createStore, select } from './state';
 
-module.exports = {
-  dom,
-  geometry,
-  gestures,
-  layout,
-  matrix,
-  state,
-};
+export const layout = { nextScene, primaryUpdate };
+export const matrix = { multiply, rotateZ, translate };
+export const state = { createStore, select };
+export const toCSS = matrixToCSS;
