+++ x-pack/plugins/canvas/public/lib/aeroelastic_kibana.js
@@ -4,12 +4,12 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import aero from './aeroelastic';
+import { layout, matrix, state } from './aeroelastic';
 
 const stores = new Map();
 
 export const aeroelastic = {
-  matrix: aero.matrix,
+  matrix,
 
   clearStores() {
     stores.clear();
@@ -20,14 +20,14 @@ export const aeroelastic = {
       throw new Error('Only a single aeroelastic store per page should exist');
     }
 
-    stores.set(page, aero.state.createStore(initialState, onChangeCallback));
+    stores.set(page, state.createStore(initialState, onChangeCallback));
 
-    const updateScene = aero.state.select((nextScene, primaryUpdate) => ({
+    const updateScene = state.select((nextScene, primaryUpdate) => ({
       shapeAdditions: nextScene.shapes,
       primaryUpdate,
       currentScene: nextScene,
       configuration: nextScene.configuration,
-    }))(aero.layout.nextScene, aero.layout.primaryUpdate);
+    }))(layout.nextScene, layout.primaryUpdate);
 
     stores.get(page).setUpdater(updateScene);
   },
